add language selection function
Modify the translateTextAsync in the original service, connect the variable of the selected language to page.xaml.cs. When the picker language is selected, the language code of the current language and the language to be translated are obtained through dictionary matching. Then clicking the translation button, two language codes are transmitted to the azure translator service to get the corresponding translation result.